+++ flavor.h
@@ -59,7 +59,7 @@ int in_send_queue(int);
 
 // calls sendfile() if !is_dev indicates that. returns 0 on success, -1 on error.
 // uses normal read/write if sendfile cannot be used. updates offset, left and copied accordingly
-ssize_t sendfile(int peer, int fd, off_t *offset, size_t n, size_t &left, size_t &copied, int);
+ssize_t sendfile(int peer, int fd, off_t *offset, size_t n, off_t &left, off_t &copied, int);
 
 int sandbox();
